requirements.txt
# ezkl
ezkl
# onnx
numpy
protobuf==3.16.0
onnx
# Keras (with PyTorch backend)
keras
torchinfo
# Jupyter notebooks
ipykernel
# Configuration management
hydra-core
